feat(azure): added azure anthropic, added backwards compat support for chat completions API, added opus 4.6 (#3145)
* feat(azure): added azure anthropic, added backwards compat support for chat completions API, added opus 4.6

* added max thinking level

* update tests

* ack comments

* update cql validation
