+++ apps/sim/app/api/tools/confluence/search-in-space/route.ts
@@ -53,7 +53,7 @@ export async function POST(request: NextRequest) {
       return NextResponse.json({ error: cloudIdValidation.error }, { status: 400 })
     }
 
-    const escapeCqlValue = (value: string) => value.replace(/"/g, '\\"')
+    const escapeCqlValue = (value: string) => value.replace(/\\/g, '\\\\').replace(/"/g, '\\"')
 
     let cql = `space = "${escapeCqlValue(spaceKey)}"`
